requirements.txt
matplotlib==3.8.1
mysql-connector-python==8.2.0
numpy==1.26.1
pandas==2.1.2
SQLAlchemy==2.0.23
urllib3==1.26.12
